==> notebooks/step_1_selecting-single-authors-titles.ipynb <==
import re
import pandas as pd

publication_file = '../metadata/DBNL 20241212 Proza 19e en 20e eeuw met subgenres.xlsx'
pub_df = pd.read_excel(publication_file)

## Preprocessing

def extract_author_id(title_id):
    if title_id.startswith('_'):
        return None
    if m := re.match(r"^(\w+\d+)(\w+\d+)", title_id):
        author_id = m.group(1)
    else:
        print('unexpected title_id format:', title)
    return author_id


def extract_year(jaar):
    jaar = jaar.strip()
    year_start, year_end, precision = None, None, None
    if m := re.match(r"^\d{4}$", jaar):
        year_start = int(jaar)
        year_end = year_start
        precision = 'exact'
    elif m := re.match(r"^(\d{4})-(\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = int(m.group(2))
        precision = 'exact'
    elif m := re.match(r"^ca\. (\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = year_start
        precision = 'circa'
    elif m := re.match(r"^ca\. (\d{4})\?$", jaar):
        year_start = int(m.group(1))
        year_end = year_start
        precision = 'circa'
    elif m := re.match(r"^na (\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = year_start
        precision = 'after'
    elif m := re.match(r"^ca\. (\d{4})-(\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = int(m.group(2))
        precision = 'circa'
    else:
        print(f"#{jaar}#")
    return year_start, year_end, precision

pub_df['author_id'] = pub_df.ti_id.apply(extract_author_id)
pub_df.author_id
pub_df['year_start'], pub_df['year_end'], pub_df['year_precision'] = zip(*pub_df.jaar.apply(extract_year))
pub_df['decade_start'] = pub_df.year_start.apply(lambda x: int(x/10) * 10)
pub_df['decade_end'] = pub_df.year_end.apply(lambda x: int(x/10) * 10)

pub_df

pub_df.genre.value_counts()

s = pub_df.ti_id.value_counts()
s = s[s > 1]
pub_df[pub_df.ti_id.isin(s.index)].sort_values('ti_id')

s = pub_df.subgenre.value_counts()
list(s[s > 10].index)

prose_fiction_genres = [
    'roman',
    'novelle(n)',
    'verhalen',
    'historische roman',
    'detective',
    'briefroman',
    'roman (populair/bestseller)',
    # 'dierenverhaal/-epos',
    # 'aforismen',
    # 'satire',
    # 'vertaling',
    # 'legende-mythe-sage',
    # 'marialegende',
    'schelmenverhaal',
    # 'vertaling: Frans / Nederlands',
    # 'bijbel / bijbeltekst(en)',
    # 'vertaling: Amerikaans-Engels / Nederlands',
    # 'heiligenleven',
    # 'toneeltekst (modern)',
    'ridderroman',
]

## Removing anonymous titles and magazines

# if `Voornaam` is `Anoniem`, the author is anonymous
pub_df['is_anonymous'] = pub_df.voornaam.apply(lambda x: True if x == 'Anoniem' else False)
print(f"number of titles with anonymous authors: {len(pub_df[pub_df.is_anonymous == True])}")

# if the `ti_id` starts with an underscore, there is no seperate author ID
pub_df['no_author_id'] = pub_df.ti_id.str.startswith('_')
print(f"number of titles with no author ID: {len(pub_df[pub_df.no_author_id == True])}")

# if the `ti_id` starts with an underscore, there is no seperate author ID
pub_df['is_magazine'] = pub_df.subgenre.str.contains('tijdschrift')
print(f"number of titles that are magazines: {len(pub_df[pub_df.is_magazine == True])}")



anonymous_ti_ids = list(pub_df[pub_df.is_anonymous == True].ti_id.unique())
authorless_ti_ids = list(pub_df[pub_df.no_author_id == True].ti_id.unique())
magazine_ti_ids = list(pub_df[pub_df.is_magazine == True].ti_id.unique())

remove_ids = list(set(anonymous_ti_ids + magazine_ti_ids + authorless_ti_ids))
print('number of title ids to remove:', len(remove_ids))
pub_df = pub_df[pub_df.ti_id.isin(remove_ids) == False]
pub_df.shape

pub_df = pub_df[(pub_df.is_anonymous == False) & (pub_df.no_author_id == False)]
s = pub_df.subgenre.value_counts()
s[s > 10]

# removing multi-author magazines/periodicals
pub_df = pub_df[pub_df.subgenre.apply(lambda x: False if 'tijdschrift' in x else True)]

pub_df.shape

## Extra check: are there multiple print editions of a work?

pub_df[['ti_id', 'druk']].drop_duplicates().ti_id.value_counts()



No, although for some titles we only have the second, third, or later print edition, for each work we only have one edition.

## Removing titles with multiple authors


ti_auth = pub_df[['ti_id', 'voornaam', 'voorvoegsel', 'achternaam']].drop_duplicates()
num_authors = ti_auth.ti_id.value_counts()
multi_authors = num_authors[num_authors > 1]
pub_df[pub_df.ti_id.isin(multi_authors.index)]

pub_df = pub_df[pub_df.ti_id.isin(multi_authors.index) == False]
pub_df.shape

## Filtering prose fiction genres

non_prose_ids = pub_df[pub_df.subgenre.isin(prose_fiction_genres) == False].ti_id.unique()

non_prose_ids.shape

### Collapsing subgenre labels

Titles with multiple subgenres (where all subgenres are in our `prose_fiction_genres` list) have one row per subgenre. We want to merge the subgenres.

pub_df['is_prose'] = pub_df.ti_id.apply(lambda x: False if x in non_prose_ids else True)
pub_df.is_prose.value_counts()

#pub_df['is_prose'] = pub_df.subgenre.apply(lambda x: True if x in prose_fiction_genres else False)
#non_prose_ids = pub_df[pub_df.is_prose == False].ti_id.unique()
prose_df = pub_df[pub_df.is_prose]

#prose_df = pub_df[pub_df.subgenre.isin(prose_fiction_genres)]
pub_df.shape, prose_df.shape


prose_df = prose_df.drop('subgenre', axis=1).drop_duplicates()
prose_df.shape

Make sure that title IDs only occur once:

prose_df.ti_id.value_counts()

pub_df = pub_df.drop('subgenre', axis=1).drop_duplicates()
pub_df.to_csv('../metadata/DBNL-single_author_pub_titles.tsv', sep='\t', index=False)

prose_df.to_csv('../metadata/DBNL-single_author_prose_titles.tsv', sep='\t', index=False)
