==> notebooks/step_2_extract_author_information.ipynb <==
# Extract Author Information

We want to obtain year of birth and year of death for authors in the DBNL dataset.

There is a database of author metadata on DBNL that I've downloaded (http://dbnl.nl/data, for login details, ask Peter).

This notebook merges the book metadata and the author metadata.

The author metadata has four columns:


| ID | Name | Volledige naam | Personalia |
| --- | --- | --- | --- |
| aafj001 | Aafjes, Bertus | [Lambertus Jacobus Johannes] | (Amsterdam 12 mei 1914 - Swolgen 22 april 1993) |
| aals010 | Aalst, George van |  | (1897 - 1925) |
| a___002 | A, Arien |  | (?(17de eeuw)) |
| a___001 | A, Herman van der |  | (Nieuw Weerdinge 1950) |

The column `Personalia` contains `DoB` and `DoD` information, where possible including geographic location and exact date of birth or death. But in many cases, either or both are unknown or only partially known. In the case of `A, Herman van der`, it is not clear if `Nieuw Weerdingen 1950` is the location and year of birth or death. To check, we should compare this year with the years of publication of titles by _Herman van der A_. If those publications are before 1950, then this is the year of death. If they are after 1950, it is the date of birth.

So we start with the book title metadata and compute the year of starting the first book and finishing the last book as the period that they were active. 

import re
import pandas as pd

publication_file = '../metadata/DBNL-single_author_prose_titles.tsv'
pub_df = pd.read_csv(publication_file, sep='\t')

We need to keep in mind that some titles have no known author (`Anoniem`) or not a single/stable (set of) of authors, such as magazines, where the contributors may change between issues.

# if `Voornaam` is `Anoniem`, the author is anonymous
pub_df['is_anonymous'] = pub_df.voornaam.apply(lambda x: True if x == 'Anoniem' else False)
print(f"number of titles with anonymous authors: {len(pub_df[pub_df.is_anonymous == True])}")

# if the `ti_id` starts with an underscore, there is no seperate author ID
pub_df['no_author_id'] = pub_df.ti_id.str.startswith('_')
print(f"number of titles with no author ID: {len(pub_df[pub_df.no_author_id == True])}")


pub_df.shape

## Extracting original print year for reprints

Some of the titles are reprints. To know where a work sits on the timeline of the authors active period, we need to know the year of the original publication. A separate spreadsheet with reprint information can be used to find the original publication year for many titles.

reprint_file = '../metadata/DBNL 20250122 Proza 19e en 20e eeuw later herdrukt.xlsx'

reprint_df = pd.read_excel(reprint_file)
reprint_df.head(2)

reprint_df['origineel._jaar'].describe()

The original publication year column contains integers, with the minimum being 1800 and the maximum 1999, so the column can be used directly. 

reprint_id_field = 'latere_uitgave.ti_id'
reprint_map = reprint_df[[reprint_id_field, 'origineel._jaar']].drop_duplicates()
s = reprint_df[reprint_id_field].value_counts()
s[s > 1]


There are reprints that have multiple rows in the spreadsheet. What is going on here?

reprint_df[reprint_df[reprint_id_field].isin(s[s > 3].index)].head(5)

Some titles in our dataset map back to multiple earlier prints. This is the case for collections of stories of an author, where the publication year of the collection maps back to the years that the individual stories were published. What we can do is map these titles to entire period, and in a later step, split these into the start year and end year.

reprint_map = (reprint_df
               .groupby(reprint_id_field)['origineel._jaar']
               .describe()[['min', 'max']]
               .apply(lambda row: f"{row['min']:.0f}-{row['max']:.0f}", axis=1)
               .rename('orig_jaar')
               .reset_index()
               .rename(columns={'latere_uitgave.ti_id': 'reprint_id'}))

reprint_map

pub_ext_df = pd.merge(pub_df, reprint_map, left_on='ti_id', right_on='reprint_id', how='left')

# pub_df and pub_ext_df should have the same number of rows
pub_df.shape, pub_ext_df.shape

print("numer of titles with mapped original year:", len(pub_ext_df[pub_ext_df.orig_jaar.notna()]))
print("numer of titles with unmapped original year:", len(pub_ext_df[pub_ext_df.orig_jaar.isna()]))

pub_ext_df['best_jaar'] = (pub_ext_df.apply(lambda row: row['jaar'] if pd.isna(row['orig_jaar']) else row['orig_jaar'], axis=1))
pub_df = pub_ext_df

## Title Identifier Check

tid_freq = pub_df.ti_id.value_counts()
tid_freq[tid_freq > 1]

## Extracting author ID and years starting and ending a publication

The title ID in most cases consist of an author ID and an identifier of one of their publications. I.e. `woen003lant04` has the author ID `woen003` for _Peter van Woensel_ and `lant04` is the specifier for _De Lantaarn voor 1800_. So the specifier has the first four letters of the first content word of the title and a running number to separate multiple publications starting with the same four letters. 

The `Jaar` column contains an indication of the year(s) of writing or publication (I'm not sure how to exactly interpret this). What I think we're interested in is the range from first to last years to establish the period of author activity. 

def extract_author_id(title_id):
    if title_id.startswith('_'):
        return None
    if m := re.match(r"^(\w+\d+)(\w+\d+)", title_id):
        author_id = m.group(1)
    else:
        print('unexpected title_id format:', title)
    return author_id


def extract_year(jaar):
    jaar = jaar.strip()
    year_start, year_end, precision = None, None, None
    if m := re.match(r"^\d{4}$", jaar):
        year_start = int(jaar)
        year_end = year_start
        precision = 'exact'
    elif m := re.match(r"^(\d{4})-(\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = int(m.group(2))
        precision = 'exact'
    elif m := re.match(r"^ca\. (\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = year_start
        precision = 'circa'
    elif m := re.match(r"^ca\. (\d{4})\?$", jaar):
        year_start = int(m.group(1))
        year_end = year_start
        precision = 'circa'
    elif m := re.match(r"^na (\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = year_start
        precision = 'after'
    elif m := re.match(r"^ca\. (\d{4})-(\d{4})$", jaar):
        year_start = int(m.group(1))
        year_end = int(m.group(2))
        precision = 'circa'
    else:
        print(f"#{jaar}#")
    return year_start, year_end, precision

pub_df['author_id'] = pub_df.ti_id.apply(extract_author_id)
pub_df.author_id
pub_df['year_start'], pub_df['year_end'], pub_df['year_precision'] = zip(*pub_df.best_jaar.apply(extract_year))

pub_df

Check the number of titles per author:

author_freq = pub_df.author_id.value_counts()
author_freq

for i in range(1, 11):
    print(f"number of authors with at least {i} titles: {len(author_freq[author_freq >= i])}")

So there are 82 authors with at least 10 titles (but this includes non-fiction publications like magazines and articles).

author_active = pd.concat([
    pub_df.groupby('author_id').year_start.min(), 
    pub_df.groupby('author_id').year_end.max()], axis=1)

author_active = (author_active
                 .rename(columns={'year_end': 'active_end', 'year_start': 'active_start'}))

author_active

Check that all authors have an explicit active period (there should be no `NaN`s):

author_active[author_active.active_start.isna()]

## Parsing the author metadata

The author data is downloaded as a large (35Mb) HTML file with multiple `table` elements.

import numpy as np

from bs4 import BeautifulSoup as Bsoup

author_file = '../data/dbnl-auteurs.html'

with open(author_file, 'rt') as fh:
    page_soup = Bsoup(fh)

def rewrite_errors(date_string):
    for error in rewrite_map:
        if re.search(r'\b' + error + r'\b', date_string):
            date_string = re.sub(r'\b' + error + r'\b', rewrite_map[error], date_string)
    return date_string
    

def parse_date_string(date_string, prefix: str):
    date_string = rewrite_errors(date_string)
    date_info = {f'{prefix}_string': date_string}
    if m := re.match(r"^(.*) (\d{1,2}) (\w+) (\d{4})", date_string):
        date_info[f'{prefix}_location'] = m.group(1)
        date_info[f'{prefix}_day'] = int(m.group(2))
        date_info[f'{prefix}_month'] = month_map[m.group(3)]
        date_info[f'{prefix}_year'] = int(m.group(4))
    elif m := re.match(r"^(.*?) (circa|ca\.) (\d{4})", date_string):
        date_info[f'{prefix}_location'] = m.group(1)
        date_info[f'{prefix}_year'] = int(m.group(3))
    elif m := re.match(r"^(.*?) (circa|ca\.) " + month_pattern + " (\d{4})", date_string):
        date_info[f'{prefix}_location'] = m.group(1)
        date_info[f'{prefix}_month'] = month_map[m.group(3)]
        date_info[f'{prefix}_year'] = int(m.group(4))
    elif m := re.match(r"^(.*?) " + month_pattern + " (\d{4})", date_string):
        date_info[f'{prefix}_location'] = m.group(1)
        date_info[f'{prefix}_month'] = month_map[m.group(2)]
        date_info[f'{prefix}_year'] = int(m.group(3))
    elif m := re.match(r"^(.*?) (\d{4})", date_string):
        date_info[f'{prefix}_location'] = m.group(1)
        date_info[f'{prefix}_year'] = int(m.group(2))
    elif m := re.match(r"^(\d{4})$", date_string):
        date_info[f'{prefix}_year'] = int(m.group(1))
    elif m := re.match(r"^\?\((\d{2})(ste|de) eeuw\)$", date_string):
        date_info[f'{prefix}_century'] = int(m.group(1))
    year = f'{prefix}_year'
    century = f'{prefix}_century'
    if year in date_info and century not in date_info:
        date_info[century] = np.ceil(date_info[year] / 100)
    # ?(20ste eeuw)
    # Rotterdam ca. 1686 
    # Correggio circa augustus 1489
    return date_info
        
def parse_personalia(row_string):
    pers_string = row_json['Personalia']
    if pers_string.startswith('(') and pers_string.endswith(')'):
        pers_string = pers_string[1:-1]
    parts = [part.strip() for part in pers_string.split(' - ')]
    pers_info = {}
    if len(parts) == 2:
        dob_string, dod_string = parts
        dob_info = parse_date_string(dob_string, 'dob')
        dod_info = parse_date_string(dod_string, 'dod')
        pers_info.update(dob_info)
        pers_info.update(dod_info)
    elif len(parts) == 1:
        dob_string = parts[0]
        dob_info = parse_date_string(dob_string, 'dob')
        if row_string['active_start'] is not None and 'dob_year' in dob_info and row_string['active_start'] < dob_info['dob_year']:
            print('switching dob to dod', pers_string)
            dob_info = parse_date_string(dob_string, 'dod')
        pers_info.update(dob_info)
    # print(pers_string)
    # print(pers_info)
    # print()
    return pers_info
        

def get_active_years(author_id, author_active):
    active_start, active_end = None, None
    if author_id in author_active.index:
        active_start = author_active.loc[author_id].active_start
        active_end = author_active.loc[author_id].active_end
    #print(author_id, active_start, active_end)
    return active_start, active_end



# Map month names to numbers
month_map = {
    "januari": 1,
    "februari": 2,
    "maart": 3,
    "april": 4,
    "mei": 5,
    "juni": 6,
    "juli": 7,
    "augustus": 8,
    "september": 9,
    "oktober": 10,
    "november": 11,
    "december": 12
}

month_pattern = "(" + "|".join(month_map.keys()) + ")"

# Map typos to correct month names (based on all lines that give an error when parsing the date string)
rewrite_map = {
    'feburari': 'februari',
    'April': 'april',
    'junu': 'juni',
    'augsustus': 'augustus',
    'augusus': 'augustus',
    'augutus': 'augustus',
    'novemer': 'november',
    'spetember': 'september',
    'decenber': 'december',
    'ecember': 'december',
    'Augustus': 'augustus',
    'Februari': 'februari',
    'Februarij': 'februari',
    'Januari': 'januari',
    'Juli': 'juli',
    'Maart': 'maart',
    'Mei': 'mei',
    'September': 'september',
    'agustus': 'augustus',
    'arpil': 'april',
    'aug': 'augustus',
    'august': 'augustus',
    'augustys': 'augustus',
    'augusuts': 'augustus',
    'dcember': 'december',
    'decembe': 'december',
    'decmber': 'december',
    'decmeber': 'december',
    'febrari': 'februari',
    'ferbruari': 'februari',
    'ferbuari': 'februari',
    'feruari': 'februari',
    'gebruari': 'februari',
    'iuli': 'juli',
    'jan': 'januari',
    'januar': 'januari',
    'janurair': 'januari',
    'janurari': 'januari',
    'jui': 'juli',
    'juil': 'juli',
    'julie': 'juli',
    'julli': 'juli',
    'maarrt': 'maart',
    'meei': 'mei',
    'nomvember': 'november',
    'nov': 'november',
    'noveber': 'november',
    'nvovember': 'november',
    'okotber': 'oktober',
    'okotober': 'oktober',
    'oktber': 'oktober',
    'oktobrt': 'oktober',
    'paril': 'april',
    'seprember': 'september',
    'septelber': 'september',
    'septemer': 'september',
    'septemner': 'september',
    'setepmber': 'september',
    'setpember': 'september',
    'sptember': 'september',
    'uni': 'juni'
}




# First, extract all table rows
rows = [tr for tr in page_soup.find_all('tr')]
print('num rows:', len(rows))

# extra the header
header_row = rows[0]
headers = [td.text for td in header_row.find_all('td')]

# turn all HTML table rows into JSON dictionaries
data_rows = rows[1:]
rows_json = []

reference_row = {}

for di, data_row in enumerate(data_rows):
    cells = [td.text.strip() for td in data_row.find_all('td')]
    
    # pad rows with missing cells by adding empty strings
    if len(cells) < len(headers):
        extra = len(headers) - len(cells)
        cells = cells + [''] * extra

    # turn row to JSON
    row_json = {header: cells[hi] for hi, header in enumerate(headers)}
    row_json['row_idx'] = di
    author_id = row_json['ID']

    if 'zie:' in author_id:
        reference_row[di] = row_json
    # skip rows in the database of author IDs that are not in title metadata
    if author_id not in author_active.index:
        #print(row_json)
        continue

    # extract the active period and add it to the JSON dictionary
    active_start, active_end = get_active_years(author_id, author_active)
    row_json['active_start'], row_json['active_end'] = active_start, active_end

    # parse the Personalia field and update the row JSON with the DoB and DOD info
    personalia = parse_personalia(row_json)
    row_json.update(personalia)

    # add the row to the list of rows
    rows_json.append(row_json)


print(f"num selected rows: {len(rows_json)}")

print('number of authors with DoB, DoD and active period:', len(rows_json))

## Removing alternative names

E.g. Albert Helman used at least 15 different pen names. In the DBNL Auteurs dataset, such alternative/pen names are indicated by the next row containing a reference (`zie:`).

prev_row = None
filtered_rows = []
for ri, row in enumerate(rows_json):
    next_row_idx = row['row_idx'] + 1
    is_reference = False
    if next_row_idx in reference_row:
        #print(ri, row)
        #print(reference_row[next_row_idx])
        is_reference = True
        #break
    row['is_reference'] = is_reference
    if 'Helman' in row['Naam']:
        #print(row)
        pass
    if 'helm003' in row['ID']:
        print(row)
        #print()
        pass
    #print(row)
    #break
    

author_df = pd.DataFrame(rows_json)
author_df

author_df = author_df[author_df.is_reference == False]
author_df = author_df.drop(['row_idx', 'is_reference'], axis=1)
author_df = author_df.rename(columns={'ID': 'author_id'})
author_df.shape

author_df[author_df.author_id.str.contains('kist00')]

Check whether we're missing any author IDs from the title metadata in the extended author info data:

author_set = set(author_df.author_id.unique())
[ai for ai in pub_df.author_id.unique() if ai not in author_set]


We're missing info on three authors. I think we'll survive. 


author_df.to_csv('../metadata/dbnl-authors.tsv.gz', index=False, sep='\t', compression='gzip')

author_df.to_excel('../metadata/dbnl-authors.xlsx', index=False)

## Combining Publication and Author information

pub_df.shape, pub_df.drop_duplicates().shape

author_df.author_id.value_counts()
author_df[author_df.author_id == 'helm003']

pub_df[pub_df.author_id == 'helm003']

[col for col in pub_df.columns if col in author_df.columns]

pub_author_df = pd.merge(pub_df, author_df, on='author_id')
pub_author_df.to_csv('../metadata/DBNL-publication_author_metadata.tsv', sep='\t', index=False)
pub_author_df.to_excel('../metadata/DBNL-publication_author_metadata.xlsx', index=False)



## Selecting Columns and Mapping Column Names

Use the columns "doc_id", "author_id" (also include author full name and surname), "birth_year", "death_year", "title", and "year_of_publication" (since it passed several filters and corrections, use "corrected_comp_year").


pub_author_df.columns

column_map = {
    'ti_id': 'doc_id',
    'titel': 'title',
    'author_id': 'author_id',
    'voornaam': 'first_name',
    'voorvoegsel': 'surname_prefix',
    'achternaam': 'surname',
    'Naam': 'full_name',
    'dob_year': 'birth_year',
    'dod_year': 'death_year',
    'year_start': 'year_of_publication'
}

pub_author_en = pub_author_df.rename(columns=column_map)[column_map.values()]
pub_author_en.to_csv('../metadata/DBNL-publication_author_metadata-english.tsv', index=False, sep='\t')

pub_author_en.doc_id.value_counts()

pub_df.shape
