==> notebooks/step_3_extract-tei-text.ipynb <==
# Extracting text from TEI files

import glob
import os

import pandas as pd

from bs4 import BeautifulSoup

meta_file = '../metadata/DBNL-publication_author_metadata-english.tsv'

random_state = 743737

title_df = pd.read_csv(meta_file, sep='\t')
#sample_df = title_df.groupby('decade_end').sample(5, random_state=random_state)
title_ids = set(title_df.doc_id.unique())

len(title_ids)

title_df

tei_dir = '../../../../Data/DBNL/xml/1800-2000/xml Proza 19e en 20e eeuw algemeen DBNL 20250122/'

tei_files = glob.glob(os.path.join(tei_dir, '*'))
tei_file_map = {tf.split('/')[-1].replace('_01.xml', ''): tf for tf in tei_files}
#tei_file_map.keys()

for ti_id in title_ids:
    if ti_id not in tei_file_map:
        print([tf_id for tf_id in tei_file_map if ti_id in tf_id])
        print(ti_id)

tei_file_map = {tf: tei_file_map[tf] for tf in tei_file_map if tf in title_ids}
len(tei_file_map)

soup_map = {}

for ti, ti_id in enumerate(tei_file_map):
    if ti_id not in tei_file_map:
        print(f"skipping {ti_id}")
        continue
    with open(tei_file_map[ti_id], 'rt') as fh:
        soup_map[ti_id] = BeautifulSoup(fh, features='xml')
    if (ti+1) > 1:
        break
        


sents = [sent for sent in doc.sents]
len(doc), len(sents[0])

selected_sents = filter_tokens(doc, 200, 200)
sum([len(sent) for sent in selected_sents])

from bs4 import BeautifulSoup
from collections import Counter
from collections import defaultdict


item_freq = Counter()
ti_item_freq = defaultdict(Counter)
chapter_item_freq = defaultdict(Counter)
div_map = defaultdict(list)

for ti, ti_id in enumerate(soup_map):
    soup = soup_map[ti_id]
    body = soup.find('body')
    divs = soup.find_all('div')
    for div in divs:
        for item in div.attrs.items():
            key = '--'.join(item)
            div_map[key].append(div)
    chapters = [div for div in soup.find_all('div') if 'type' in div.attrs and div.attrs['type'] == 'chapter']
    item_freq.update(['--'.join(item) for div in body.find_all('div') for item in div.attrs.items()])
    ti_item_freq[ti_id].update(['--'.join(item) for div in body.find_all('div') for item in div.attrs.items()])
    for ci, chapter in enumerate(chapters):
        ch_id = f"{ti_id}-chapter-{ci:0>3}"
    
        chapter_item_freq[ch_id].update(['--'.join(item) for div in chapter.find_all('div') for item in div.attrs.items()])
    children = [child for child in body if child.name is not None]
    if (ti+1) > 500:
        break

item_freq

All `div` elements with attribute `ebook-type` or `wpg-type` are not part of the main prose text.

def extract_attr_items(div):
    return ['--'.join(item) for item in div.attrs.items()]


item_rel_freq = Counter()

for ti, ti_id in enumerate(soup_map):
    soup = soup_map[ti_id]
    body = soup.find('body')
    divs = soup.find_all('div')
    for parent_div in divs:
        parent_items = extract_attr_items(parent_div)
        for child_div in parent_div.find_all('div'):
            child_items = extract_attr_items(child_div)
            item_rel_freq.update([(parent_item, child_item) for parent_item in parent_items for child_item in child_items])

for item in sorted(item_rel_freq):
    print(item, item_rel_freq[item])

### Non-Dutch elements

for key in div_map:
    #print(key)
    #if key != 'type--chapter':
    #    continue
    #print(Counter([item for div in div_map[key] for item in div.attrs.items()]))
    for div in div_map[key]:
        if 'type' in div.attrs and div.attrs['type'] == 'chapter':
            if 'ebook-type' in div.attrs:
                print(div)
        if 'lang' in div.attrs and div.attrs['lang'] == 'srn':
            print(div)
    #print(div.attrs)

Issues:

- non-prose elements
- non-Dutch elements

import pandas as pd

ti_item_freq
vocab = list(item_freq.keys())
rows = []

for ti_id in ti_item_freq:
    row = [ti_id] + [ti_item_freq[ti_id][attr] if attr in ti_item_freq[ti_id] else 0 for attr in vocab]
    rows.append(row)


body_df = pd.DataFrame(rows, columns=['ti_id'] + vocab)
body_df.sum()

import pandas as pd

vocab = list(item_freq.keys())
rows = []

for ch_id in chapter_item_freq:
    row = [ch_id] + [chapter_item_freq[ch_id][attr] if attr in chapter_item_freq[ch_id] else 0 for attr in vocab]
    rows.append(row)


chapter_df = pd.DataFrame(rows, columns=['ch_id'] + vocab)
chapter_df.sum()

#df['type--chapter'].min()
#df.sum()

chapter_df.columns

body = soup.find('body')
children = [child for child in body if child.name is not None]
children[0]
[child.attrs for child in children]
item_freq.update([item for div in body.find_all('div') for item in div.attrs.items()])


import spacy


nlp_full = spacy.load("nl_core_news_md")

# Load the pipeline without the entity recognizer
nlp_small = spacy.load("nl_core_news_md", exclude=["ner", "parser"])
nlp_small.add_pipe("sentencizer")

%%time
doc_full = nlp_full(dutch_prose_text)


%%time
doc_small = nlp_small(dutch_prose_text)


nlp_small.max_length = 10_000_000
nlp_small.max_length

stats_rows = []

out_file = '../data/dbnl-chunks-tokenised-skip_first_200.tsv.gz'

headers = ['title_id', 'chunk_id', 'term', 'pos', 'term_freq']

with gzip.open(out_file, 'wt') as fh_out:
    header_string = '\t'.join(headers)
    fh_out.write(f"{header_string}\n")
    for ti, ti_id in enumerate(tei_file_map):
        with open(tei_file_map[ti_id], 'rt') as fh_in:
            soup = BeautifulSoup(fh_in, features='xml')
            dutch_prose_text = extract_dutch_prose_text(soup)
            doc = nlp_small(dutch_prose_text)
            chunks = chunk_doc(doc, skip_first=200)
            for row in make_chunk_rows(ti_id, chunks):
                row_string = '\t'.join([str(cell) for cell in row])
                fh_out.write(f"{row_string}\n")
        print(f"{ti+1} books parsed")

## Filter Output by Available Titles

(2025-04-09: this only needs to be done because we have filtered more problematic (non-fiction) titles from the dataset)

import datetime

today = datetime.date.today().isoformat()

import gzip


orig_file = '../data/dbnl-chunks-tokenised-skip_first_200.tsv.gz'
new_file = f'../data/dbnl-chunks-tokenised-skip_first_200-version-{today}.tsv.gz'

len(title_ids)

with gzip.open(orig_file, 'rt') as fh_in:
    with gzip.open(new_file, 'wt') as fh_out:
        header_line = next(fh_in)
        fh_out.write(header_line)
        for line in fh_in:
            title_id, *rest = line.split('\t')
            if title_id in title_ids:
                fh_out.write(line)
            else:
                continue
                print(title_id, title_id in title_ids, rest)
                break

chapters_texts = extract_chapter_texts(soup)
for chapter_texts in chapters_texts:
    words = [word for text in chapter_texts for word in re.split(r"\s+", text)]
    print(len(words))

df = Counter()

prev_chunk = None
terms = set()
with gzip.open(out_file, 'rt') as fh:
    headers = next(fh).strip('\n').split('\t')
    for line in fh:
        cells = line.strip('\n').split('\t')
        row = {header: cells[hi] for hi, header in enumerate(headers)}
        if prev_chunk and prev_chunk != row['chunk_id']:
            df.update(terms)
            terms = set()
            
        #df.update([cells[2]])
        terms.add(row['term'])
        #print(cells)
        prev_chunk = row['chunk_id']


df.most_common(10)


len(df), sum(df.values())

num_chunks = 29260
max_df = num_chunks / 2
max_df

rank = 0
for term, freq in df.most_common():
    rank += 1
    if freq < max_df:
        print(rank, term, freq)
        break
    

from frog import Frog, FrogOptions


frog = Frog(FrogOptions(parser=False, ner=False, chunking=False, morph=False))


#output = frog.process_raw("Dit is een test")
#print("RAW OUTPUT=",output)
output = frog.process("Dit is nog een test.")
#print("PARSED OUTPUT=",output)
len(output)
for token in output:
    print(token)

import time

start = time.time()
output = frog.process(dutch_prose_text)
step = time.time()
took = step - start
took

start = time.time()
doc = nlp_small(dutch_prose_text)
step = time.time()
took = step - start
took

len(output)

print(' '.join([t['text'] for t in output[:200]]))

print(' '.join([t['chunker'] for t in output[:200]]))

output[197]

frog_1 = Frog(FrogOptions(parser=False, ner=False, chunking=False, morph=False))

start = time.time()
output = frog_1.process(dutch_prose_text)
step = time.time()
took = step - start
took

frog_4 = Frog(FrogOptions(parser=False, ner=False, chunking=False, morph=False, numThreads=b"4"))

start = time.time()
output = frog_4.process(dutch_prose_text)
step = time.time()
took = step - start
took

## Testing lemmatisation of older spelling

import glob
import os

from bs4 import BeautifulSoup

from tei_parsing import extract_dutch_prose_text

tei_dir = '../../../../Data/DBNL/xml/1800-2000/xml Proza 19e en 20e eeuw algemeen DBNL 20250122/'
tei_files = glob.glob(os.path.join(tei_dir, '*'))
tei_file = [t for t in tei_files if 'kist001' in t][0]
with open(tei_file, 'rt') as fh:
    soup = BeautifulSoup(fh, features='xml')
    dutch_prose_text = extract_dutch_prose_text(soup)

from frog import Frog, FrogOptions

frog = Frog(FrogOptions(parser=False, ner=False, chunking=False, morph=False))

len(dutch_prose_text)
output = frog.process(dutch_prose_text)

sent_old = "Vijf-en-twintig jaren zijn verloopen sedert tallooze tegenstrijdigheden in Chateaubriand's letterkundige nalatenschap."
sent_modern = "Vijf-en-twintig jaren zijn verlopen sedert talloze tegenstrijdigheden in Chateaubriand's letterkundige nalatenschap."

sent_out = frog.process(sent_old)
sent_out

import spacy


#nlp_small = spacy.load("nl_core_news_md", exclude=["ner", "parser"])
#nlp_small.add_pipe("sentencizer")
doc = nlp_small(sent_old)
print(f'------------\nOLD: {sent_old}\n------------')
for spacy_sent in doc.sents:
    for token in spacy_sent:
        print([token.text, token.lemma_, token.pos_])


doc = nlp_small(sent_modern)
print(f'------------\nMODERN: {sent_modern}\n------------')
for spacy_sent in doc.sents:
    for token in spacy_sent:
        print([token.text, token.lemma_, token.pos_])




def has_long_vowel(term):
    long_vowels = ['ee', 'aa', 'oo', 'uu']
    return any([lv in term for lv in long_vowels])


[t['text'] for ti, t in enumerate(output[:800]) if has_long_vowel(t['text'])]
